refactor: simplify Homebrew service responsibilities
