+++ CaskHub/Services/LocalHomebrewService.swift
@@ -9,141 +9,6 @@ import AppKit
 import Foundation
 import Observation
 
-// MARK: - Public Types
-
-struct LocalCaskInstallation: Hashable, Identifiable {
-    let token: String
-    let installedVersion: String
-    let installedAt: Date?
-    let appBundleNames: [String]
-
-    /// Brew still lists this cask, but its app was removed outside Homebrew
-    /// (or its install receipt is gone) — opens and upgrades are doomed.
-    let isZombie: Bool
-
-    init(
-        token: String,
-        installedVersion: String,
-        installedAt: Date?,
-        appBundleNames: [String],
-        isZombie: Bool = false
-    ) {
-        self.token = token
-        self.installedVersion = installedVersion
-        self.installedAt = installedAt
-        self.appBundleNames = appBundleNames
-        self.isZombie = isZombie
-    }
-
-    var id: String {
-        token
-    }
-}
-
-enum CaskAction: Equatable {
-    case opening
-    case installing
-    case adopting
-    case updating
-    case uninstalling
-    case queued
-
-    var inProgressLabel: String {
-        switch self {
-        case .opening: return "Opening…"
-        case .installing: return "Installing…"
-        case .adopting: return "Adopting…"
-        case .updating: return "Updating…"
-        case .uninstalling: return "Uninstalling…"
-        case .queued: return "Queued…"
-        }
-    }
-}
-
-enum LocalHomebrewError: LocalizedError {
-    case brewBinaryNotFound
-    case appBundleNotFound(token: String)
-    case brewCommandFailed(args: [String], exitCode: Int32, stderr: String)
-
-    /// Machine states (no Homebrew installed), not bugs — never worth a Sentry event.
-    var isEnvironmental: Bool {
-        if case .brewBinaryNotFound = self { return true }
-        return false
-    }
-
-    /// Coarse classes for Sentry grouping — one issue per way brew fails, not per cask.
-    static func failureClass(stderr: String) -> String {
-        if isStrandedApp(stderr: stderr) { return "stranded-caskroom-app" }
-        if stderr.contains("is not there") { return "missing-artifact-source" }
-        if stderr.contains("different from the one being installed") { return "adopt-version-mismatch" }
-        if stderr.contains("Operation not permitted") { return "permission-denied" }
-        if stderr.contains("No Cask with this name exists") || stderr.contains("No casks found") {
-            return "unknown-cask"
-        }
-        if stderr.contains("is not installed") { return "not-installed" }
-        if stderr.contains("reports different checksum") || stderr.contains("SHA256 mismatch") {
-            return "checksum-mismatch"
-        }
-        if stderr.contains("already an App at") { return "app-conflict" }
-        return "uncategorized"
-    }
-
-    /// A previous interrupted operation parked the real .app inside the Caskroom
-    /// version directory; every upgrade then fails until the copy is cleared.
-    static func isStrandedApp(stderr: String) -> Bool {
-        stderr.contains("already an App at") && stderr.contains("Caskroom")
-    }
-
-    var errorDescription: String? {
-        switch self {
-        case .brewBinaryNotFound:
-            return "Couldn't locate the brew binary. Is Homebrew installed?"
-        case let .appBundleNotFound(token):
-            return "Couldn't find an installed app for \(token)."
-        case let .brewCommandFailed(args, code, stderr):
-            let cmd = (["brew"] + args).joined(separator: " ")
-            let trimmed = stderr.trimmingCharacters(in: .whitespacesAndNewlines)
-            if Self.isAppManagementDenial(stderr: trimmed) {
-                return "macOS blocked CaskHub from modifying apps on your Mac. "
-                    + "Enable CaskHub under System Settings → Privacy & Security → "
-                    + "App Management, then try again."
-            }
-            if Self.isStrandedApp(stderr: trimmed) {
-                return "A previous update left an old copy of the app inside Homebrew's "
-                    + "records, and Homebrew refuses every upgrade until it's cleared. "
-                    + "Repair removes the leftover copy and reinstalls the app fresh — "
-                    + "your settings and data are kept."
-            }
-            if Self.isAdoptMismatch(args: args, stderr: trimmed) {
-                return "Your installed copy doesn't match the version Homebrew has on record, "
-                    + "so it can't be adopted as-is. You can replace it with Homebrew's copy "
-                    + "instead — your settings and data are kept."
-            }
-            if trimmed.contains("reports different checksum") || trimmed.contains("SHA256 mismatch") {
-                return "The download doesn't match the checksum Homebrew has on record — "
-                    + "the developer likely replaced the release file after it was published. "
-                    + "This isn't a problem with your Mac; Homebrew refuses mismatched downloads "
-                    + "for security. Try again in a day or two once the cask is updated."
-            }
-            return trimmed.isEmpty
-                ? "`\(cmd)` failed (exit \(code))."
-                : "`\(cmd)` failed (exit \(code)): \(trimmed)"
-        }
-    }
-
-    /// The App Management (TCC) permission gating modification of other apps' bundles.
-    static func isAppManagementDenial(stderr: String) -> Bool {
-        stderr.contains("Operation not permitted")
-    }
-
-    /// `brew install --adopt` refuses when the on-disk app differs from the cask's version.
-    static func isAdoptMismatch(args: [String], stderr: String) -> Bool {
-        args.contains("--adopt")
-            && (stderr.contains("different from the one being installed")
-                || stderr.contains("already an App at"))
-    }
-}
-
 // MARK: - LocalHomebrewService
 
 @MainActor
@@ -157,7 +22,7 @@ final class LocalHomebrewService {
     /// Executable names found in common install locations (~/.local/bin, /usr/local/bin, …).
     var externalBinaryNames: Set<String> = []
 
-    private(set) var adoptReplaceOffers: Set<String> = []
+    var adoptReplaceOffers: Set<String> = []
 
     /// Casks wedged by a stranded app copy inside the Caskroom; repair =
     /// clear brew's records, then reinstall fresh.
@@ -166,7 +31,7 @@ final class LocalHomebrewService {
     private(set) var appManagementDenials: Set<String> = []
 
     /// Adoptions waiting for the App Management permission; value = retry with `--force`.
-    private(set) var permissionRequests: [String: Bool] = [:]
+    var permissionRequests: [String: Bool] = [:]
 
     /// Test seam — the real probe hits TCC via the filesystem.
     @ObservationIgnored var permissionProbe: @Sendable () -> AppManagementPermission.Status
@@ -192,7 +57,7 @@ final class LocalHomebrewService {
 
     @ObservationIgnored private var runningProcesses: [String: Process] = [:]
 
-    private(set) var actionErrors: [String: String] = [:]
+    var actionErrors: [String: String] = [:]
 
     private(set) var isUpdatingAll = false
 
@@ -205,9 +70,9 @@ final class LocalHomebrewService {
     /// Include self-updating casks (`auto_updates: true`) in updates, via `brew upgrade --greedy`.
     private(set) var greedyUpdates: Bool
 
-    private let fileManager: FileManager
+    let fileManager: FileManager
     private let defaults: UserDefaults
-    private let applicationDirectories: [URL]
+    let applicationDirectories: [URL]
 
     private static let greedyKey = "greedyUpdates"
 
@@ -285,23 +150,6 @@ final class LocalHomebrewService {
         lastRefresh = .now
     }
 
-    func isInstalled(token: String) -> Bool {
-        installedCasks[token] != nil
-    }
-
-    /// The cask isn't brew-managed, but its app already sits in /Applications.
-    func isAdoptable(_ cask: Cask) -> Bool {
-        installedCasks[cask.token] == nil
-            && cask.appArtifactNames.contains(where: externalAppNames.contains)
-    }
-
-    /// A CLI cask whose tool is on the device via some other installer
-    /// (e.g. claude-code's native install script). Detected, not managed.
-    func isExternalCLI(_ cask: Cask) -> Bool {
-        installedCasks[cask.token] == nil
-            && cask.binaryArtifactNames.contains(where: externalBinaryNames.contains)
-    }
-
     func clearError(for token: String) {
         actionErrors[token] = nil
         adoptReplaceOffers.remove(token)
@@ -309,82 +157,7 @@ final class LocalHomebrewService {
         appManagementDenials.remove(token)
     }
 
-    func isOutdated(token: String, remoteVersion: String) -> Bool {
-        guard let installation = installedCasks[token], !installation.isZombie else { return false }
-        return Self.comparableVersion(installation.installedVersion)
-            != Self.comparableVersion(remoteVersion)
-    }
-
-    private nonisolated static func comparableVersion(_ version: String) -> Substring {
-        version.prefix { $0 != "," && $0 != "_" }
-    }
-
-    func hasAvailableUpdate(token: String, remoteVersion: String, autoUpdates: Bool?) -> Bool {
-        (greedyUpdates || autoUpdates != true) && isOutdated(token: token, remoteVersion: remoteVersion)
-    }
-
     // MARK: - Actions
-
-    /// The scan's zombie verdict cross-checked against the *current* cask
-    /// artifacts: auto-updating apps can rename their bundle on disk
-    /// (Codex.app → ChatGPT.app), stranding the install-time receipt and
-    /// symlink while the app lives on under its new name. Deletion is only
-    /// offered when the apps the cask declares today are verifiably gone too.
-    func isZombie(_ cask: Cask) -> Bool {
-        guard let installation = installedCasks[cask.token], installation.isZombie,
-              !cask.appArtifactNames.isEmpty
-        else { return false }
-        return existingBundleURL(named: cask.appArtifactNames) == nil
-    }
-
-    /// Opens via the receipt's bundle names, falling back to the cask's current
-    /// artifact names — receipts go stale when apps rename themselves.
-    func open(_ cask: Cask) {
-        actionErrors[cask.token] = nil
-        let receiptNames = installedCasks[cask.token]?.appBundleNames ?? []
-        openBundle(named: receiptNames + cask.appArtifactNames, token: cask.token)
-    }
-
-    func openApp(token: String) {
-        actionErrors[token] = nil
-
-        guard let installation = installedCasks[token] else {
-            actionErrors[token] = LocalHomebrewError.appBundleNotFound(token: token).errorDescription
-            return
-        }
-        openBundle(named: installation.appBundleNames, token: token)
-    }
-
-    /// Launches a not-yet-adopted app straight from its on-disk bundle.
-    func openExternalApp(cask: Cask) {
-        actionErrors[cask.token] = nil
-        openBundle(named: cask.appArtifactNames, token: cask.token)
-    }
-
-    /// Version of a not-yet-adopted app, read from its bundle's Info.plist.
-    func externalAppVersion(for cask: Cask) -> String? {
-        guard installedCasks[cask.token] == nil,
-              let appURL = existingBundleURL(named: cask.appArtifactNames),
-              let info = Bundle(url: appURL)?.infoDictionary
-        else { return nil }
-        return info["CFBundleShortVersionString"] as? String
-            ?? info["CFBundleVersion"] as? String
-    }
-
-    private func openBundle(named names: [String], token: String) {
-        guard let appURL = existingBundleURL(named: names) else {
-            actionErrors[token] = LocalHomebrewError.appBundleNotFound(token: token).errorDescription
-            return
-        }
-        appLauncher(appURL)
-    }
-
-    private func existingBundleURL(named names: [String]) -> URL? {
-        names.flatMap { name in
-            applicationDirectories.map { $0.appendingPathComponent(name) }
-        }
-        .first { fileManager.fileExists(atPath: $0.path) }
-    }
 
     func install(token: String) async throws {
         try await runMutation(.installing, token: token, args: ["install", "--cask", token])
@@ -482,7 +255,7 @@ final class LocalHomebrewService {
         return Self.strandedCopyExists(in: caskroom, token: token, fileManager: fileManager)
     }
 
-    private func runMutation(_ action: CaskAction, token: String, args: [String]) async throws {
+    func runMutation(_ action: CaskAction, token: String, args: [String]) async throws {
         guard inFlightActions[token] == nil || inFlightActions[token] == .queued else { return }
         inFlightActions[token] = action
         actionErrors[token] = nil
@@ -544,7 +317,8 @@ final class LocalHomebrewService {
         let collector = BrewOutputCollector()
         collector.attach(to: process, pipe: pipe) { [weak self] text in
             guard text.contains("==> Installing Cask") else { return }
-            Task { @MainActor in self?.cancellableDownloads.remove(token) }
+            guard let self else { return }
+            Task { @MainActor in self.cancellableDownloads.remove(token) }
         }
 
         try process.run()
@@ -560,99 +334,5 @@ final class LocalHomebrewService {
                 stderr: outputTail.components(separatedBy: "\n").suffix(6).joined(separator: "\n")
             )
         }
-    }
-}
-
-// MARK: - Adoption
-
-extension LocalHomebrewService {
-    /// Preflight before `--adopt`: a declared binary missing from the on-disk
-    /// bundle makes brew fail *after* it has moved the app aside — and brew's
-    /// rollback then deletes the app entirely. Refuse locally and offer the
-    /// safe replace path instead.
-    func adopt(_ cask: Cask, bypassPermissionCheck: Bool = false) async throws {
-        if let missing = adoptBlockedByMissingBinary(cask) {
-            adoptReplaceOffers.insert(cask.token)
-            actionErrors[cask.token] = "Your installed copy of \(cask.displayName) is missing "
-                + "a component Homebrew's version includes (\(missing)), so it can't be "
-                + "adopted as-is. You can replace it with Homebrew's copy instead — "
-                + "your settings and data are kept."
-            return
-        }
-        try await adopt(token: cask.token, bypassPermissionCheck: bypassPermissionCheck)
-    }
-
-    /// Only paths inside the app bundle can be preflighted; staged-path binaries
-    /// don't exist until brew downloads the cask, so they're skipped.
-    func adoptBlockedByMissingBinary(_ cask: Cask) -> String? {
-        guard let appURL = existingBundleURL(named: cask.appArtifactNames) else { return nil }
-        let marker = "/\(appURL.lastPathComponent)/"
-        for path in cask.binarySourcePaths {
-            guard let range = path.range(of: marker) else { continue }
-            let resolved = appURL.appendingPathComponent(String(path[range.upperBound...]))
-            if !fileManager.fileExists(atPath: resolved.path) {
-                return URL(fileURLWithPath: path).lastPathComponent
-            }
-        }
-        return nil
-    }
-
-    func adopt(token: String, bypassPermissionCheck: Bool = false) async throws {
-        if !bypassPermissionCheck, await !permissionAllowsAdoption() {
-            permissionRequests[token] = false
-            return
-        }
-        do {
-            try await runMutation(.adopting, token: token, args: ["install", "--cask", token, "--adopt"])
-        } catch {
-            if case let LocalHomebrewError.brewCommandFailed(args, _, stderr) = error,
-               LocalHomebrewError.isAdoptMismatch(args: args, stderr: stderr),
-               !LocalHomebrewError.isStrandedApp(stderr: stderr) {
-                adoptReplaceOffers.insert(token)
-            }
-            throw error
-        }
-    }
-
-    /// Fallback when `--adopt` refuses: replace the on-disk app with Homebrew's copy.
-    /// Needs App Management even more than adopt — brew deletes the existing app,
-    /// and a TCC-denied delete makes brew escalate to a scary sudo password prompt.
-    func adoptReplacing(token: String, bypassPermissionCheck: Bool = false) async throws {
-        if !bypassPermissionCheck, await !permissionAllowsAdoption() {
-            permissionRequests[token] = true
-            return
-        }
-        adoptReplaceOffers.remove(token)
-        try await runMutation(.adopting, token: token, args: ["install", "--cask", token, "--force"])
-    }
-
-    func cancelPermissionRequest(token: String) {
-        permissionRequests[token] = nil
-    }
-
-    /// Called when the app becomes active: if the user granted App Management while
-    /// away (in System Settings), finish the adoptions that were waiting on it.
-    func resumePendingAdoptions() {
-        guard !permissionRequests.isEmpty else { return }
-        Task {
-            guard await permissionAllowsAdoption() else { return }
-            let pending = permissionRequests
-            permissionRequests.removeAll()
-            for (token, useForce) in pending {
-                if useForce {
-                    try? await adoptReplacing(token: token, bypassPermissionCheck: true)
-                } else {
-                    try? await adopt(token: token, bypassPermissionCheck: true)
-                }
-            }
-        }
-    }
-
-    /// `.unknown` passes through — only a confirmed denial blocks, so a failed
-    /// probe can never lock the user out of adopting.
-    private func permissionAllowsAdoption() async -> Bool {
-        let probe = permissionProbe
-        let status = await Task.detached(priority: .userInitiated) { probe() }.value
-        return status != .denied
     }
 }
